requirements.txt
requests>=2,<3
numpy>=1.24,<3
pandas>=2,<3
pillow>=10,<12
opencv-python>=4,<5
matplotlib>=3,<4
seaborn>=0.13,<1
scipy>=1.11,<2
scikit-learn>=1,<2
scikit-image>=0.22,<1
h5py>=3,<4
tqdm>=4,<5
python-dotenv>=1,<2
astropy>=6,<8
PyYAML>=6,<7
torch>=2,<3
torchvision>=0.17,<1
tensorflow>=2,<3
shap>=0.44,<1
notebook>=7,<8
ipykernel>=6,<7
